// servicecloud/cti/twilio/bui/web/cti_messaging_addin/scripts/model/message.ts
/* * *******************************************************************************************
 *  This file is part of the Oracle Service Cloud Accelerator Reference Integration set 
 *  published by Oracle under the Universal Permissive License (UPL), Version 1.0
 *  Copyright (c) 2017 Oracle and/or its affiliates. All rights reserved.
 ***********************************************************************************************
 *  Accelerator Package: Telephony and SMS Accelerator for Twilio
 *  link: http://www.oracle.com/technetwork/indexes/samplecode/accelerator-osvc-2525361.html
 *  OSvC release: 17D (November 2017)
 *  reference: 161212-000059
 *  date: Monday Oct 30 13:4:53 UTC 2017
 *  revision: rnw-17-11-fixes-releases
 * 
 *  SHA1: $Id: 79532489fd91d43cef3310af2b3344b2c79a66e8 $
 * *********************************************************************************************
 *  File: message.ts
 * ****************************************************************************************** */

import {Contact} from "./contact";
export class Message {
    key: string;
    message: string;
    contact: Contact;
    incidentWorkspace: any;
}